TSK-1730: Moved all Builder classes to package builder
