+++ lib/taskana-core/src/main/java/pro/taskana/classification/internal/builder/ClassificationBuilder.java
@@ -1,4 +1,4 @@
-package pro.taskana.classification.internal;
+package pro.taskana.classification.internal.builder;
 
 import java.security.PrivilegedActionException;
 import java.security.PrivilegedExceptionAction;
