+++ lib/taskana-core/src/main/java/pro/taskana/task/internal/builder/TaskCommentBuilder.java
@@ -1,4 +1,4 @@
-package pro.taskana.task.internal;
+package pro.taskana.task.internal.builder;
 
 import java.security.PrivilegedActionException;
 import java.security.PrivilegedExceptionAction;
