+++ lib/taskana-core/src/main/java/pro/taskana/workbasket/internal/builder/WorkbasketTestImpl.java
@@ -1,4 +1,4 @@
-package pro.taskana.workbasket.internal;
+package pro.taskana.workbasket.internal.builder;
 
 import java.time.Instant;
