+++ lib/taskana-core/src/test/java/acceptance/builder/TaskBuilderTest.java
@@ -4,10 +4,10 @@ import static acceptance.DefaultTestEntities.defaultTestObjectReference;
 import static acceptance.DefaultTestEntities.defaultTestWorkbasket;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
-import static pro.taskana.classification.internal.ClassificationBuilder.newClassification;
+import static pro.taskana.classification.internal.builder.ClassificationBuilder.newClassification;
 import static pro.taskana.common.internal.util.CheckedSupplier.wrap;
 import static pro.taskana.task.internal.builder.TaskBuilder.newTask;
-import static pro.taskana.workbasket.internal.WorkbasketAccessItemBuilder.newWorkbasketAccessItem;
+import static pro.taskana.workbasket.internal.builder.WorkbasketAccessItemBuilder.newWorkbasketAccessItem;
 
 import java.time.Instant;
 import java.util.List;
@@ -34,7 +34,7 @@ import pro.taskana.task.api.TaskService;
 import pro.taskana.task.api.TaskState;
 import pro.taskana.task.api.models.Attachment;
 import pro.taskana.task.api.models.Task;
-import pro.taskana.task.internal.TaskAttachmentBuilder;
+import pro.taskana.task.internal.builder.TaskAttachmentBuilder;
 import pro.taskana.task.internal.builder.TaskBuilder;
 import pro.taskana.task.internal.models.TaskImpl;
 import pro.taskana.workbasket.api.WorkbasketPermission;
